refactor: Add helper methosds from season stats to module
